+++ deal/src/main/java/ru/litlmayn/deal/mappers/FinishRegistrationToClientMapper.java
@@ -1,11 +1,11 @@
-package ru.litlmayn.mappers;
+package ru.litlmayn.deal.mappers;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.factory.Mappers;
 import ru.litlmayn.api.dto.FinishRegistrationRequestDto;
-import ru.litlmayn.models.Client;
+import ru.litlmayn.deal.models.Client;
 
 @Mapper(componentModel = "sprig")
 public interface FinishRegistrationToClientMapper {
