+++ deal/src/main/java/ru/litlmayn/deal/models/Client.java
@@ -1,4 +1,4 @@
-package ru.litlmayn.models;
+package ru.litlmayn.deal.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
