+++ deal/src/main/java/ru/litlmayn/deal/repositories/ClientRepository.java
@@ -1,8 +1,7 @@
-package ru.litlmayn.repositories;
+package ru.litlmayn.deal.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import ru.litlmayn.models.Client;
-import ru.litlmayn.models.Statement;
+import ru.litlmayn.deal.models.Client;
 
 import java.util.Optional;
 import java.util.UUID;
