+++ deal/src/main/java/ru/litlmayn/deal/services/.DS_Store
