+++ deal/src/main/java/ru/litlmayn/deal/services/impl/.DS_Store
