+++ deal/src/test/java/ru/litlmayn/mappers/CreditDtoToCreditMapperTest.java
@@ -3,6 +3,7 @@ package ru.litlmayn.mappers;
 import org.junit.jupiter.api.Test;
 import ru.litlmayn.api.dto.CreditDto;
 import ru.litlmayn.api.dto.PaymentScheduleElementDto;
+import ru.litlmayn.deal.mappers.CreditDtoToCreditMapper;
 import ru.litlmayn.models.Credit;
 
 import java.math.BigDecimal;
