+++ deal/src/test/java/ru/litlmayn/mappers/LoanOfferRequestToClientMapperTest.java
@@ -2,6 +2,7 @@ package ru.litlmayn.mappers;
 
 import org.junit.jupiter.api.Test;
 import ru.litlmayn.api.dto.LoanStatementRequestDto;
+import ru.litlmayn.deal.mappers.LoanOfferRequestToClientMapper;
 import ru.litlmayn.models.Client;
 
 import java.math.BigDecimal;
